+++ client/cl_main.c
@@ -1142,9 +1142,9 @@ void CL_Disconnect (void)
 	final[0] = clc_stringcmd;
 //	strncpy ((char *)final+1, "disconnect");
 	Q_strncpyz ((char *)final+1, "disconnect", sizeof(final)-1);
-	Netchan_Transmit (&cls.netchan, strlen(final), final);
-	Netchan_Transmit (&cls.netchan, strlen(final), final);
-	Netchan_Transmit (&cls.netchan, strlen(final), final);
+	Netchan_Transmit (&cls.netchan, (int)strlen(final), final);
+	Netchan_Transmit (&cls.netchan, (int)strlen(final), final);
+	Netchan_Transmit (&cls.netchan, (int)strlen(final), final);
 
 	CL_ClearState ();
 
@@ -1209,7 +1209,7 @@ void CL_Packet_f (void)
 	out = send+4;
 	send[0] = send[1] = send[2] = send[3] = (char)0xff;
 
-	l = strlen (in);
+	l = (int)strlen (in);
 	for (i=0 ; i<l ; i++)
 	{
 		if (in[i] == '\\' && in[i+1] == 'n')
