+++ qcommon/cmd.c
@@ -98,14 +98,14 @@ void Cbuf_AddText (char *text)
 {
 	int		l;
 	
-	l = strlen (text);
+	l = (int)strlen (text);
 
 	if (cmd_text.cursize + l >= cmd_text.maxsize)
 	{
 		Com_Printf ("Cbuf_AddText: overflow\n");
 		return;
 	}
-	SZ_Write (&cmd_text, text, strlen (text));
+	SZ_Write (&cmd_text, text, (int)strlen (text));
 }
 
 
@@ -315,7 +315,7 @@ qboolean Cbuf_AddLateCommands (void)
 	argc = COM_Argc();
 	for (i=1 ; i<argc ; i++)
 	{
-		s += strlen (COM_Argv(i)) + 1;
+		s += (int)strlen (COM_Argv(i)) + 1;
 	}
 	if (!s)
 		return false;
@@ -578,7 +578,7 @@ char *Cmd_MacroExpandString (char *text)
 	inquote = false;
 	scan = text;
 
-	len = strlen (scan);
+	len = (int)strlen (scan);
 	if (len >= MAX_STRING_CHARS)
 	{
 		Com_Printf ("Line exceeded %i chars, discarded.\n", MAX_STRING_CHARS);
@@ -603,7 +603,7 @@ char *Cmd_MacroExpandString (char *text)
 	
 		token = Cvar_VariableString (token);
 
-		j = strlen(token);
+		j = (int)strlen(token);
 		len += j;
 		if (len >= MAX_STRING_CHARS)
 		{
@@ -700,7 +700,7 @@ void Cmd_TokenizeString (char *text, qboolean macroExpand)
 			cmd_args[sizeof(cmd_args)-1] = 0; 
 
 			// strip off any trailing whitespace
-			l = strlen(cmd_args) - 1;
+			l = (int)strlen(cmd_args) - 1;
 			for ( ; l >= 0 ; l--)
 				if (cmd_args[l] <= ' ')
 					cmd_args[l] = 0;
@@ -819,7 +819,7 @@ Cmd_CompleteCommand
 	int				len;
 	cmdalias_t		*a;
 	
-	len = strlen(partial);
+	len = (int)strlen(partial);
 	
 	if (!len)
 		return NULL;
@@ -854,7 +854,7 @@ char *Cmd_CompleteCommand (char *partial)
 	char			*pmatch[1024];
 	qboolean		diff = false;
 	
-	len = strlen(partial);
+	len = (int)strlen(partial);
 	
 	if (!len)
 		return NULL;
