+++ ui/ui_options_controls.c
@@ -192,7 +192,7 @@ void Options_Controls_MenuInit ( void )
 	s_options_controls_header.generic.type		= MTYPE_SEPARATOR;
 	s_options_controls_header.generic.textSize	= MENU_HEADER_FONT_SIZE;
 	s_options_controls_header.generic.name		= "Controls";
-	s_options_controls_header.generic.x			= MENU_HEADER_FONT_SIZE/2 * strlen(s_options_controls_header.generic.name);
+	s_options_controls_header.generic.x			= MENU_HEADER_FONT_SIZE/2 * (int)strlen(s_options_controls_header.generic.name);
 	s_options_controls_header.generic.y			= -2*MENU_LINE_SIZE;	// 0
 
 	s_options_controls_sensitivity_slider.generic.type		= MTYPE_SLIDER;
